+++ .storybook/manager.ts
@@ -1,5 +1,5 @@
 import {addons} from '@storybook/addons';
-import {themeLight} from './theme';
+import {themeLight} from './theme/light';
 
 addons.setConfig({
     theme: themeLight,
