+++ .storybook/theme/dark.ts
@@ -1,0 +1,5 @@
+import {create} from '@storybook/theming';
+
+export const themeDark = create({
+    base: 'dark',
+});
