+++ .storybook/theme/index.ts
@@ -1,0 +1,7 @@
+import {themeLight} from './light';
+import {themeDark} from './dark';
+
+export const themes = {
+    light: themeLight,
+    dark: themeDark,
+};
